+++ src/main/java/com/codetaylor/mc/athenaeum/util/BlockRegistrationHelper.java
@@ -1,4 +1,4 @@
-package com.codetaylor.mc.athenaeum.helper;
+package com.codetaylor.mc.athenaeum.util;
 
 import com.codetaylor.mc.athenaeum.spi.IBlockColored;
 import com.codetaylor.mc.athenaeum.spi.IBlockVariant;